+++ src/main/java/sync/slamtalk/mate/entity/RecruitmentStatusType.java
@@ -1,0 +1,5 @@
+package sync.slamtalk.mate.entity;
+
+public enum RecruitmentStatusType {
+    RECRUITING, COMPLETED, CANCEL
+}
