+++ src/main/java/sync/slamtalk/mate/entity/SkillLevelType.java
@@ -1,0 +1,5 @@
+package sync.slamtalk.mate.entity;
+
+public enum SkillLevelType {
+    BEGINNER, INTERMEDIATE, MASTER, IRRELAVANT
+}
