+++ src/main/java/sync/slamtalk/mate/repository/MatePostRepository.java
@@ -2,7 +2,7 @@ package sync.slamtalk.mate.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import sync.slamtalk.mate.domain.MatePost;
+import sync.slamtalk.mate.entity.MatePost;
 
 @Repository
 public interface MatePostRepository extends JpaRepository<MatePost, Long> {
